refactor: use immutable types from @ariadnejs/types package
Updated all call_analysis modules to use TypeDiscovery, TypeInfo,
CallAnalysisResult, and SimpleRange from the @ariadnejs/types package
instead of defining them locally. This ensures consistency with the
project's immutable architecture and reduces code duplication.

- Import TypeDiscovery and CallAnalysisResult from types package
- Import SimpleRange from types package for range utilities
- Import TypeInfo for type tracking
- Re-export types for convenience in index.ts
